search: RG-only (published_rg) filter in Searcher.prepare
Previously `published_rg` was passed as a params arg in multiple
places around the app but was not in the search fields whitelist
so it was ignored.
When the field was added to the whitelist it caused ES parsing
problems (conflict betw ES query_string and bool filters?).
This commit adds `published_rg` filter to all searches on Page
objects as part of Searcher.prepare.